pyproject.toml
[tool.poetry]
name = "templa"
version = "0.1.0"
description = ""
authors = ["yuxki <yuxki.ryoshi@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.16"
jinja2 = "3.1.2"
pyyaml = "6.0"

[tool.poetry.group.dev.dependencies]
tox = "^4.5.1"
pre-commit = "^3.3.2"
mypy = "^1.3.0"
dodgy = "^0.2.1"
flake8 = "^6.0.0"
types-pyyaml = "^6.0.12.9"
black = "^23.3.0"
pytest = "^7.3.1"


[tool.poetry.group.docs.dependencies]
sphinx = "^7.0.1"
furo = "^2023.5.20"
pydantic = "^1.10.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.messages_control]
ignore-paths = [
  "tests",
  "docs",
]
disable = [
  # intentionaly turned off
  "import-error",
  "too-few-public-methods",
  "pointless-string-statement",
  "invalid-name",
  "too-many-ancestors",
  "duplicate-code",
  "useless-import-alias",

  # pylint type "C": convention, for programming standard violation↲
  "missing-class-docstring",
  "missing-function-docstring",
  "missing-module-docstring",
]

[tool.mypy]
files = ["src/templa", "tests"]
python_version = 3.11
show_error_codes = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_any_generics = true
no_implicit_optional = true
local_partial_types = true
no_implicit_reexport = true
strict_equality = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
